Install socket.io and necessary dependencies
First, `socket.io` was installed and added as a dependency, along with `socket.io-client`
so that the client can communicate via `socket.io`. In addition, the `buffer-utils` and
`utf-8-validate` modules were manually added because of issues with webpack and
next.js that came from a recent-ish update. More information can be found at this
(issue)[https://github.com/vercel/next.js/issues/49334].