Add middleware.ts file to redirect users when not logged in
The `middleware.ts` is a file that should sit in the top level directory
that nextjs uses automatically to perform some redirects/callbacks. At
the moment, we use the `next-auth` default middleware, and choose
`/sockettest` as a route that we want to prevent users from visiting
without being signed in.