Add prisma adapter to next-auth
The next-auth prisma adapter is a built-in integration for `next-auth.js`,
which automatically sends data to whatever database prisma is configured to
interact with. This allows us to automatically store user, account, and
session information. For more information about this, and the expected schema,
see this: [adapters](https://authjs.dev/reference/adapter/prisma)